- Implemented static factory methods CoefFunction::Generate() for real /   complex valued scalar / vector and tensor functions. These methods   investigate the expression to be set and generate and initialize the   appropriate CoefFunction class, e.g. if the expression evaluates to a   constant, a CoefFunctionConst is generated, otherwise a   CoefFunctionExpression.
  The mechanism could be extended to allow for
  expression, that only change "sometimes", e.g. that depend on
  time/space, where we could trigger the re-calculation of the internal
  matrix by a call-back function.

- Fixed compile issues with debug version of intel composer 12.1.0
  20111011, build 7. This was mostly due to "possible loose of
  significant bits" and some warnings on "meaningless" constructs within
  boosts. We suppress theses messages. A detailed list with explanations
  can be found in cmake_modules/compiler.cmake.

- Fixed some constructs in CFS itself, where no explicit cast from int
  to double was performed.

+++ source/Forms/IntScheme.cc
@@ -916,7 +916,7 @@ void IntScheme::GetIntPoints( Elem::ShapeType elemType,
     int i, dtbl, m;
 
     // calculate number of points required according to p = 2n-1
-    int n = ((order + 1) / 2.0) + 0.5;
+    int n = int(((order + 1) / 2.0) + 0.5);
 
     // not sure, for what m is used for
     m = (n+1)>>1;
